Update: Clean code, separate PacketEvent, add some bypass before 1.0 (part 2) (#76)

+++ src/main/java/net/minusmc/minusbounce/features/module/modules/movement/speeds/watchdog/WatchdogSpeed.kt
@@ -15,7 +15,7 @@ class WatchdogSpeed: SpeedMode("Watchdog", SpeedType.WATCHDOG) {
             } else {
                 MovementUtils.strafe(0.44f)
             }
-            mc.thePlayer.motionY = MovementUtils.getJumpBoostModifier(0.42F, true)
+            mc.thePlayer.motionY = MovementUtils.getJumpBoostModifier(0.42F, true).toDouble()
         }
     }
